Fix cast in debug print

+++ lib/network-protocol/TCP.cpp
@@ -220,7 +220,7 @@ void NetworkProtocolTCP::status_server(NetworkStatus *status)
     {
         status->connected = server->hasClient();
         status->error = error;
-        Debug_printf("TCP::status_server C:%d E:%d\n", status->connected, status->error);
+        Debug_printf("TCP::status_server C:%d E:%d\n", status->connected, (int) status->error);
     }
 }
